Add 'metadata' option to include @metadata field in output
Just like the rubydebug codec, being able to include the @metadata
field in the output of the json_lines codec can be useful, for
example for debugging or test tools that prefer machine-readable
output.

The new option is immediately marked as deprecated to discourage
its use and to make it easier to replace it with something better
once we figure out what that would be.

Since the Event class's constructor mutates the input hash I had
to refactor some existing tests to send in a clone of the hash
since it'll otherwise end up without a @metadata field.